move types to appropriate place
