+++ nifi-nar-bundles/nifi-framework-bundle/nifi-framework/nifi-web/nifi-web-frontend/src/main/nifi/src/app/pages/settings/state/reporting-tasks/index.ts
@@ -45,6 +45,19 @@ export interface ConfigureReportingTaskRequest {
     postUpdateNavigation?: string[];
 }
 
+export interface ConfigureReportingTaskSuccess {
+    id: string;
+    reportingTask: ReportingTaskEntity;
+    postUpdateNavigation?: string[];
+}
+
+export interface ConfigureReportingTaskRequest {
+    id: string;
+    uri: string;
+    payload: any;
+    postUpdateNavigation?: string[];
+}
+
 export interface EditReportingTaskDialogRequest {
     id: string;
     reportingTask: ReportingTaskEntity;
