+++ nifi-nar-bundles/nifi-framework-bundle/nifi-framework/nifi-web/nifi-web-frontend/src/main/nifi/src/app/pages/settings/state/reporting-tasks/reporting-tasks.actions.ts
@@ -27,9 +27,10 @@ import {
     StartReportingTaskRequest,
     StartReportingTaskSuccess,
     StopReportingTaskRequest,
-    StopReportingTaskSuccess
+    StopReportingTaskSuccess,
+    ConfigureReportingTaskRequest,
+    ConfigureReportingTaskSuccess
 } from './index';
-import { ConfigureReportingTaskRequest, ConfigureReportingTaskSuccess } from '../management-controller-services';
 
 export const resetReportingTasksState = createAction('[Reporting Tasks] Reset Reporting Tasks State');
